+++ packages/database/src/index.ts
@@ -16,6 +16,7 @@ export * from "./saved-resources.js";
 export * from "./templates.js";
 export * from "./display-settings.js";
 export * from "./lifecycle.js";
+export * from "./export.js";
 export * from "./pool.js";
 
 export type DatabaseHealthCode =